Partner space

+++ app/Http/Controllers/CurrencyController.php
@@ -11,7 +11,6 @@ class CurrencyController extends Controller
     public function __construct()
     {
          $this->middleware('auth:staff');
-         $this->middleware('CanRead:currency'); //->except('index','create');
     }
     /**
      * Get a validator for an incoming registration request.
@@ -22,9 +21,8 @@ class CurrencyController extends Controller
     protected function validateRequest(Request $request)
     {
         $request->validate([
-            'name' => 'required|unique:Currencies,name',
-            'symbole' => 'required|unique:Currencies,symbole',
-            'country_id' => 'required',
+            'name' => 'required|unique:currencies,name',
+            'symbole' => 'required|unique:currencies,symbole',
         ]);
     }
 
@@ -60,13 +58,34 @@ class CurrencyController extends Controller
     {
         $this->validateRequest($request);
 
+
+        $trashedCurrency = Currency::onlyTrashed()->where('name', $request->name)->first();
+        if(isset($trashedCurrency))
+        {
+            return redirect('currencies/'.$trashedCurrency->id);
+        }
+
         $currency = new Currency();
         $currency->name = $request->name;
         $currency->symbole = $request->symbole;
-        $currency->country_id = $request->country_id;
         $currency->save(); 
         
-        return redirect('currencies');
+        return $currency;
+    }
+     /**
+     * 
+     */
+    public function storeWithRedirect(Request $request) {
+        $currency = self::store($request);
+        return redirect('currencies/'.$currency->id);
+    }
+
+    /**
+     * 
+     */
+    public function storeAndNew(Request $request) {
+        $currency = self::store($request);
+        return redirect('currencies/create');
     }
 
     /**
@@ -77,9 +96,9 @@ class CurrencyController extends Controller
      */
     public function show($Currency)
     {
-        
+        // All the informations are shown on currency index
         $data['Currency'] = Currency::find($Currency);
-        return 1;
+        return view();
     }
     
     /**
@@ -105,10 +124,10 @@ class CurrencyController extends Controller
     public function update(Request $request, $currency)
     {
         $request->validate([
-            'name' => 'required|unique:currencies,name,'.$currency,
-            'symbole' => 'required|unique:currencies,symbole,'.$currency,
-            'country_id' => 'required',
-        ]);
+            'name' => 'required|unique:currencies,name,'.$currency.',id',
+            'symbole' => 'required|unique:currencies,symbole,'.$currency.',id',
+            'country_id' => 'required|unique:currencies,country_id,'.$currency.',id',
+            ]);
         
         $currency = Currency::find($currency);
         $currency->name = $request->name;
@@ -130,8 +149,93 @@ class CurrencyController extends Controller
     {
         // récupérer photo
         $currency = Currency::findOrFail($currency);
-       $currency->delete();
-       return redirect('currencies');
+        if(isset($currency->products[0]))
+        {
+            $messages['error'] = 'Currency can\'t be deleted it has products !!';
+            return  redirect('currencies')
+                        ->with('messages',$messages);
+         
+        }
+        
+        $currency->delete();
+        $messages['success'] = 'Currency deleted successfuly !!';
+        return  redirect('currencies')
+                    ->with('messages',$messages);
 
     }
+
+     /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\Detail  $detail
+     * @return \Illuminate\Http\Response
+     */
+    public function multiDestroy(Request $request)
+    {
+
+        $request->validate([
+            'currencies' => 'required',
+        ]);
+        $e=$s=0;
+        $messages = [];
+        
+        foreach($request->currencies as $Currency)
+        {
+
+            $currency = Currency::findOrFail($Currency);
+    
+            if(!isset($currency->products[0]) && !isset($currency->countries[0])) 
+            {
+                $s++;
+                $currency->delete();
+                $messages['success'] = $s. ($s == 1 ? ' currency' :' currencies') .' has been deleted successfuly';
+            }
+            else 
+            {
+                $e++;
+                $messages['error'] = $e . ($e == 1 ? ' currency' : ' currencies') . ' can\'t be deleted it has a relation with products';
+            }
+            return redirect('currencies')->with('messages', $messages);
+
+        }
+    }
+ 
+
+    public function restore($Currency)
+    {
+        $currency = Currency::onlyTrashed()->where('id', $Currency)->first();
+        $currency->restore();
+        $messages['success'] = 'Currency has been restored successfuly !!';
+        return redirect('currencies')->with('messages',$messages);
+    }
+
+    public function multiRestore(Request $request)
+    {
+        $request->validate([
+            'details' => 'required',
+        ]);
+        $s=0;
+        $messages = [];
+        foreach ($request->currencies as  $Currency)
+        {
+            $currency = Currency::onlyTrashed()->where('id', $Currency)->first();
+            $currency->restore();
+            $s++;
+            $messages['success'] = $s. ($s == 1 ? ' currency' :' currencies') .' has been restored successfuly';
+        }
+        return redirect('currencies')->with('messages',$messages);
+    }
+
+    /**
+     * Displaying the Trash page
+     * 
+     * 
+     * @return \Illuminate\Http\Response
+     */
+    public function trash()
+    {
+        $data['currency'] = Currency::onlyTrashed()->get();
+        return view('currencies.backoffice.staff.trash', $data);
+    }
+
 }
