+++ app/Http/Controllers/ParticularClientController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use Hash;
+use Password;
 use App\ParticularClient;
 use App\Country;
 use App\Picture;
@@ -16,7 +18,7 @@ class ParticularClientController extends Controller
 {
     public function __construct()
     {
-         //$this->middleware('auth:particular_clients');
+        //$this->middleware('auth:particular_clients, staff');
     }
 
     protected function validateRequest(Request $request)
@@ -26,11 +28,23 @@ class ParticularClientController extends Controller
             'first_name' => 'required',
             'last_name' => 'required',
             'email' => 'required|email|unique:particular_clients,email',
-            'gender' => 'required',
+            'gender_id' => 'required',
             'birthday' => 'required|date',
             'password' => 'required|min:6',
         ]);
     }
+
+    /**
+     * Display a trashed listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function unactive()
+    {
+        $data['particulars'] = ParticularClient::onlyTrashed()->get();
+        return view('clients.backoffice.staff.unactive',$data);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -38,14 +52,14 @@ class ParticularClientController extends Controller
      */
     public function index()
     {
-        $data['particularClients'] = ParticularClient::all();
-        //return $data;
-        return view('particular_clients.backoffice.index',$data);
+        $data['particulars'] = ParticularClient::all();
+        //  return $data['particulars'][0];
+        return view('clients.backoffice.staff.index',$data);
     }
     
-    public function dashboard()
+    public function home()
     {
-        return view('system.backoffice.particular_clients.dashboard');
+        return view('system.frontend.www.index');
     }
     /**
      * Show the form for creating a new resource.
@@ -67,23 +81,28 @@ class ParticularClientController extends Controller
     public function store(Request $request)
     {
         $this->validateRequest($request);
+
         $AddressController = new AddressController();
         $AddressController->validateRequest($request);
+
+        $PhoneController = new PhoneController();
+        $PhoneController->validateRequest($request);
         
         $request->validate([
             'numbers.0' => 'required|numeric|unique:phones,number,'.$request->phone_id[0],
-            'numbers.1' => 'sometimes|numeric|unique:phones,number,'.$request->phone_id[1],
+            'numbers.1' => 'nullable|numeric|unique:phones,number|digits:9',
+            'numbers.1' => 'nullable|numeric|unique:phones,number|digits:9',
             'code_country.0' => 'required',
             'code_country.1' => 'required',
-        ]); 
-          //return $request->numbers;
+        ]);
+        
         $is_register_to_newsletter = ($request->is_register_to_newsletter=='on') ? 1 : 0;
         $particularClient = new ParticularClient();
         $particularClient->name = $request->name;
         $particularClient->first_name = $request->first_name;
         $particularClient->last_name = $request->last_name;
         $particularClient->email = $request->email;
-        $particularClient->gender_id = $request->gender;
+        $particularClient->gender_id_id = $request->gender_id;
         $particularClient->birthday = date('Y-m-d H:i:s',strtotime($request->birthday));
         $particularClient->password = bcrypt($request->password);
         $particularClient->is_register_to_newsletter = $is_register_to_newsletter;
@@ -98,7 +117,7 @@ class ParticularClientController extends Controller
                 $phone->number = $number;
                 $phone->type = "phone";
                 $phone->country_id = $request->code_country[$key];
-                $phone->phoneable_type = 'particularClient';
+                $phone->phoneable_type = 'particular_client';
                 $phone->phoneable_id = $particularClient->id;
                 $phone->save();
             }
@@ -111,7 +130,7 @@ class ParticularClientController extends Controller
             $phone->number = $request->fax_number;
             $phone->type = "fix";
             $phone->country_id = $request->code_country[2];
-            $phone->phoneable_type = 'particularClient';
+            $phone->phoneable_type = 'particular_client';
             $phone->phoneable_id = $particularClient->id;
             $phone->save();
         }
@@ -121,12 +140,13 @@ class ParticularClientController extends Controller
         {
             $address = new Address();
             $address->address = $request->address;
-            $address->address_tow = $request->address_tow;
+            $address->address_two = $request->address_two;
             $address->full_name = $request->full_name;
             $address->zip_code = $request->zip_code;
             $address->country_id = $request->country_id;
             $address->city = $request->city;
-            $address->addressable_type = 'particularClient';
+            $address->is_default = true;
+            $address->addressable_type = 'particular_client';
             $address->addressable_id = $particularClient->id;
             $address->save();
             
@@ -135,15 +155,15 @@ class ParticularClientController extends Controller
         {
             $picture = Picture::create([
                 'name' => time().'.'.$request->file('path')->extension(),
-                'tag' => "particularClient",
+                'tag' => "particular_client_avatar",
                 'path' => $request->path->store('images/particular_clients', 'public'),
                 'extension' => $request->path->extension(),
-                'pictureable_type' => 'particularClient',
+                'pictureable_type' => 'particular_client',
                 'pictureable_id' => $particularClient->id,
                 ]);
                 
             }
-            return redirect('particular-clients');
+            return redirect('clients/particular');
 
        
     }
@@ -154,11 +174,22 @@ class ParticularClientController extends Controller
      * @param  \App\particular_clients  $particular_clients
      * @return \Illuminate\Http\Response
      */
-    public function show($particular_clients)
+    public function show($client)
     {
-        $data['particularClient'] = ParticularClient::where('name',$particular_clients)->first();
-        // return $data['particular_clients']->phones[0]->country; 
-        return view('particular_clients.backoffice.show',$data);
+        $data['client'] = ParticularClient::withTrashed()->where('name',$client)->first();
+        $activities = $data['client']->logs
+                                            ->where('created_at','>=' ,date('Y-m-d h:i:s',strtotime("-1 week") ))
+                                            ->groupBy(function($item){return $item->created_at->format('d-M-y');});
+
+        
+        $data['activities'] = [];
+        foreach($activities as $key => $activitys)
+        {
+            $data['activities'][$key]= $activitys->sortByDesc('created_at');
+        }
+
+        krsort($data['activities']);
+        return view('clients.backoffice.staff.show',$data);
     }
 
     /**
@@ -175,7 +206,6 @@ class ParticularClientController extends Controller
         //$data['particular_clients'] = ParticularClient::find(Auth::guard('particular_clients')->user()->id);
 
         return view('system.backoffice.particular_clients.profile',$data);
-       
     }
 
     /**
@@ -190,7 +220,6 @@ class ParticularClientController extends Controller
         $data['particularClient'] = ParticularClient::where('name',$particular_clients)->first();
         return view('particular_clients.backoffice.edit',$data);
         return $data;
-      
     }
 
     /**
@@ -206,7 +235,7 @@ class ParticularClientController extends Controller
             'first_name' => 'required',
             'last_name' => 'required',
             'email' => 'required|email|unique:particular_clients,email,'.$particulaClient,
-            'gender' => 'required',
+            'gender_id' => 'required',
             'birthday' => 'required|date',
         ]);
         $AddressController = new AddressController();
@@ -214,7 +243,6 @@ class ParticularClientController extends Controller
         
         $request->validate([
             'numbers.0' => 'required|numeric|unique:phones,number,'.$request->phone_id[0],
-            'numbers.1' => 'sometimes|numeric|unique:phones,number,'.$request->phone_id[1],
             'code_country.0' => 'required',
             'code_country.1' => 'required',
         ]);      
@@ -223,25 +251,24 @@ class ParticularClientController extends Controller
         $particularClient->first_name = $request->first_name;
         $particularClient->last_name = $request->last_name;
         $particularClient->email = $request->email;
-        $particularClient->gender_id = $request->gender;
+        $particularClient->gender_id = $request->gender_id;
         $particularClient->birthday = date('Y-m-d H:i:s',strtotime($request->birthday));
         $particularClient->is_register_to_newsletter = $is_register_to_newsletter;
         $particularClient->save();
         
         $address = $particularClient->address;
         $address->address = $request->address;
-        $address->address_tow = $request->address_tow;
+        $address->address_two = $request->address_two;
         $address->full_name = $request->full_name;
         $address->zip_code = $request->zip_code;
         $address->country_id = $request->country_id;
         $address->city = $request->city;
         $address->save();
-             
         if($request->hasFile('path')) 
         {
             $picture = $particularClient->picture;
             $picture->name = time().'.'.$request->file('path')->extension();
-            $picture->tag = "particularClient";
+            $picture->tag = "particular_client_avatar";
             $picture->path = $request->path->store('images/particular_clients', 'public');
             $picture->extension = $request->path->extension();
             $picture->save();
@@ -252,8 +279,16 @@ class ParticularClientController extends Controller
         {
             if($number != null)
             {
-               
-                $phone = Phone::find($request->phone_id[$key]);
+                $phone = Phone::where('id', $request->phone_id[$key])
+                                                                ->whereIn('type', ['phone', 'fix'])
+                                                                ->where('phoneable_id', $particularClient->id)
+                                                                ->first();
+                if($phone == null)
+                {
+                    $phone = new Phone();
+                    $phone->phoneable_id = $particularClient->id;
+                    $phone->phoneable_type = 'particular_client';
+                }
                 $phone->number = $number;
                 $phone->type = "phone";
                 $phone->country_id = $request->code_country[$key];
@@ -263,14 +298,23 @@ class ParticularClientController extends Controller
 
         if($request->fax_number)
         {
-            
+            $fax = Phone::where('id', $request->fax_id)
+                                ->where('type', 'fax')
+                                ->where('phoneable_id', $partner->id)
+                                ->first();
+            if($fax == null)
+            {
+                $fax = new Phone();
+                $phone->phoneable_id = $partner->id;
+                $phone->phoneable_type = 'particular_client';
+            }
             $phone = Phone::find($request->fax_id);
             $phone->number = $request->fax_number;
             $phone->type = "fix";
             $phone->country_id = $request->code_country[2];
             $phone->save();
         }
-        return redirect('particular-clients');
+        return redirect('clients/particular');
     }
 
 
@@ -283,8 +327,52 @@ class ParticularClientController extends Controller
     public function destroy($particular)
     {
         $particulaClient = ParticularClient::where('name', $particular)->first();
+        if($this->stuckParticularClient($particulaClient))
+        {
+            return redirect()
+                            ->back()
+                            ->with(
+                                'error',
+                                'Particular can\'t be deleted it has unsolved orders/markets !!'
+                            );
+        }
         $particulaClient->delete();
-        return redirect('particular-clients');
+        return redirect($this->redirectURL(url()->current(), $particular))
+                                ->with(
+                                    'success',
+                                    'Particular has been deleted successfuly !!'
+                                );
+    }
+
+    public function stuckParticularClient($particularClient)
+    {
+        $orderStatus = $particularClient->orders()->whereIn('status', ['in_progress', 'finished'])->get();
+        //$marketStatus = $client->markets()->whereIn('status', ['in_progress', 'finished'])->get();
+        isset($orderStatus[0]) ? $stuck = true : $stuck = false;
+        return $stuck;
+    }
+
+    public function redirectURL($url, $particular)
+    {
+        $destroy_url = str_after($url, '/'.$particular);
+        if($destroy_url == '/desactivate')
+        {
+            return str_before($url, $particular).''.$particular.'/security';
+        }
+        return str_before($url, '/'.$particular);
+    }
+
+    public function multiDestroy(Request $request)
+    {
+        foreach($request->particular_clients as $particular_client)
+        {
+            $particular_client = Business::where('name', $particular_client)->first();
+            if(!$this->stuckParticularClient($particular_client))
+            {
+                $particular_client->delete();
+            }
+        }
+        return redirect('clients/particular');
     }
 
     /**
@@ -307,7 +395,7 @@ class ParticularClientController extends Controller
      * Desable particular_clientss account.
      *
      * @param  \Iluminate\Http\Request $request
-     * @param  \App\particular_clients  $particular_clients
+     * @param  \App\ParticularClient  $particularClient
      * @param  \App\Guest  $session_id
      * @return \Illuminate\Http\Response
      */
@@ -315,5 +403,71 @@ class ParticularClientController extends Controller
     {
         DB::table('sessions')->where('id', $session_id)->delete();
         return redirect(str_before(url()->current(), '.com').'.com/security');
+    }
+
+    public function restore($particular_client)
+    {
+        $particular_client = ParticularClient::onlyTrashed()->where('name', $particular_client)->first();
+        $particular_client->restore();
+        $messages['success'] = 'Particular client has been restored successfuly !!';
+        return redirect('clients/particular')->with('messages',$messages);
+    }
+
+    public function multiRestore()
+    {
+        $request->validate([
+            'particular_clients' => 'required',
+        ]);
+        $s=0;
+        $messages = [];
+        foreach ($request->particular_clients as  $particular_client)
+        {
+            if($particular_client != null)
+            {
+                $particular_client = Business::onlyTrashed()->where('name', $particular_client)->first();
+                $particular_client->restore();
+                $s++;
+                $messages['success'] = $s. ($s == 1 ? ' particular client' :' particular clients') .' has been restored successfuly';
+            }
+        }
+        return redirect('clients/particular')->with('messages',$messages);
+    }
+
+    public function reset(Request $request, $client)
+    {
+        $password_communicated = ($request->password_communicated=='on') ? 1 : 0;
+        if($password_communicated)
+        {
+            $password = Hash::make($request->password);
+            $client =ParticularClient::where('name', $client)->first();
+            if($client)
+            {
+                $client->password = $password;
+                $client->save();
+                $messages['success'] = 'Password reset has been done successfuly !!';
+            }
+            else
+            {
+                $messages['error'] = 'Client member doesn\'t exist !!';
+            }
+        }
+        else
+        {
+            $messages['error'] = 'Please communicate the password !!';
+        }
+        return redirect()
+                        ->back()
+                        ->with('messages', $messages);
+    }
+    public function sendResetLinkEmail($client) {
+        $client = ParticularClient::where('name', $client)->first();
+        $token = Password::getRepository()->create($client);
+        return 1;
+
+        $client->sendPasswordResetNotification($token);
+
+        $messages['success'] = 'Password reset has been sent successfuly !!';
+
+        return back()->with('messages', $messages);
     }
 }
